+++ CHANGELOG.md
@@ -11,13 +11,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Use MacOSX matplotlib backend on macOS, still use TkAgg on Linux
-
 ### Added
 
 ### Removed
 
 ### Deprecated
+
+## [24.4.0] - 2024-05-30
+
+### Fixed
+
+- Fixed another odd libcxx issue between miniconda and conda-forge (involving `tuple`)
+- Fix issue with ipython (see https://github.com/ipython/ipython/issues/14260) by uninstalling
+  and re-installing `prompt_toolkit`
+
+### Changed
+
+- Updated to Miniconda 24.4.0-0.
+- Use MacOSX matplotlib backend on macOS and TkAgg on Linux
+  - Update sed command to make this change
+
+### Added
+
+- Explicit Conda Packages
+  - earthaccess
+- Explicit Pip Packages
+  - pyephem
+- Added ephem example
+
+### Removed
+
+- Removed old sed command on `numpy` as file being altered no longer exists
 
 ## [24.1.2] - 2024-04-09